refactor(onboarding): isolate installer process execution
